move pages out of components folder
