+++ src/pages/UserPage.jsx
@@ -1,57 +1,57 @@
 import React, { useState, useEffect } from "react";
-import { useLocation } from "react-router-dom";
-import { getQueryValue } from "../../util/url";
-import { Box, Typography, Grid } from "@mui/material";
-import PlanCard from "../Common/PlanCard";
-import { fetchPlans } from "../../util";
-import WelcomeMessage from "../Common/search/WelcomeMessage";
-import EmptyResultMessage from "../Common/search/EmptyResultMessage";
-import Pagination from "../Common/Pagination";
-import PlanCardSkeleton from "../Common/PlanCardSkeleton";
-import { useAuth } from "../../context/AuthContext";
+import { useParams, useLocation } from "react-router-dom";
+import { getQueryValue } from "@/util/url";
+import { Box, Typography, Grid, Avatar } from "@mui/material";
+import PlanCard from "@/Components/Common/PlanCard";
+import { fetchPlans } from "@/util";
+import Pagination from "@/Components/Common/Pagination";
+import PlanCardSkeleton from "@/Components/Common/PlanCardSkeleton";
+import { useAuth } from "@/context/AuthContext";
 
 const PAGE_SIZE = 8;
 
-const SearchPage = () => {
+const UserPage = () => {
+  const { userId } = useParams();
   const location = useLocation();
-  const [searchQuery, setSearchQuery] = useState(
-    getQueryValue(location.search, "q"),
-  );
   let page = getQueryValue(location.search, "page") || "1";
   const [isLoading, setIsLoading] = useState(true);
   const [plans, setPlans] = useState([]);
   const [pagesCount, setPagesCount] = useState(0);
+  const [user, setUser] = useState({});
   const [error, setError] = useState(null);
   const { token } = useAuth();
 
   useEffect(() => {
     page = getQueryValue(location.search, "page") || "1";
-    setSearchQuery(getQueryValue(location.search, "q"));
   }, [location.search]);
 
   useEffect(() => {
     (async () => {
       setIsLoading(true);
       const params = new URLSearchParams(location.search);
-      params.set("search", searchQuery);
       params.set("page", page);
       params.set("size", PAGE_SIZE);
       const paramsString = params.toString();
 
       try {
-        const res = await fetchPlans(`plans?${paramsString}`, token, setError);
-        setPlans(res?.items || []);
-        setPagesCount(res?.pagesCount || 0);
+        const res = await fetchPlans(
+          `plans/user/${userId}?${paramsString}`,
+          token,
+          setError,
+        );
+        const { plans: userPlans, ...userDetails } = res;
+        console.log(userPlans, userDetails);
+        setPlans(userPlans.items || []);
+        setUser(userDetails || {});
+        setPagesCount(userPlans.pagesCount || 0);
         setIsLoading(false);
       } catch (error) {
         console.log("Error fetching data:", error);
       }
     })();
-  }, [searchQuery, page]);
+  }, [page]);
 
-  if (searchQuery === "") return <WelcomeMessage />;
-
-  if (!isLoading && plans.length === 0) return <EmptyResultMessage />;
+  if (!isLoading && plans.length === 0) return <>There is no plans.</>;
 
   return (
     <Box
@@ -62,9 +62,24 @@ const SearchPage = () => {
     >
       <Typography
         variant="h5"
-        sx={{ marginBottom: "16px", fontWeight: "bold" }}
+        sx={{
+          mb: 4,
+          fontWeight: "bold",
+          display: "flex",
+          alignItems: "center",
+          gap: 1,
+        }}
       >
-        Search result for <u>{searchQuery}</u>:
+        {user.name && (
+          <>
+            <Avatar
+              alt={user.name}
+              src={user.imageURL}
+              sx={{ bgcolor: "#4CAF50" }}
+            />
+            {user.name}
+          </>
+        )}
       </Typography>
       <Box
         sx={{
@@ -95,4 +110,4 @@ const SearchPage = () => {
   );
 };
 
-export default SearchPage;
+export default UserPage;
