+++ src/pages/dashboard/MyPlans.jsx
@@ -1,17 +1,17 @@
 import React, { useState, useEffect } from "react";
-import { useNavigate } from "react-router-dom";
-import { removeBookmark, getBookmarks } from "../../../util/dashboard";
-import PlanCard from "../../Common/PlanCard";
-import PlanCardSkeleton from "../../Common/PlanCardSkeleton";
+import { Link, useNavigate } from "react-router-dom";
+import { deletePlan, getMyPlans } from "@/util/dashboard";
+import PlanCard from "@/Components/Common/PlanCard";
+import PlanCardSkeleton from "@/Components/Common/PlanCardSkeleton";
 import { Box, Grid, Button, Typography } from "@mui/material";
-import AlertDialog from "../../Common/AlertDialog";
-import { useAuth } from "../../../context/AuthContext";
-import { getQueryValue } from "../../../util/url";
-import Pagination from "../../Common/Pagination";
+import AlertDialog from "@/Components/Common/AlertDialog";
+import { useAuth } from "@/context/AuthContext";
+import { getQueryValue } from "@/util/url";
+import Pagination from "@/Components/Common/Pagination";
 
 const PAGE_SIZE = 9;
 
-function Bookmarked() {
+function MyPlans() {
   const [plans, setPlans] = useState([]);
   const [error, setError] = useState(null);
   const [alertOpen, setAlertOpen] = useState(false);
@@ -37,7 +37,7 @@ function Bookmarked() {
       params.set("size", PAGE_SIZE);
       const paramsString = params.toString();
 
-      const data = await getBookmarks(paramsString, token, setError);
+      const data = await getMyPlans(paramsString, token, setError);
       if (!data) return setIsLoading(false);
 
       setPlans(data.items || []);
@@ -47,7 +47,7 @@ function Bookmarked() {
   }, [page]);
 
   const handleRemovePlan = async () => {
-    const result = await removeBookmark(token, selectedPlanToRemove, setError);
+    const result = await deletePlan(token, selectedPlanToRemove, setError);
 
     if (!result) return;
 
@@ -70,8 +70,16 @@ function Bookmarked() {
     return (
       <Box sx={{ textAlign: "center", marginTop: 4 }}>
         <Typography variant="h5" color="textSecondary">
-          Bookmarked Plans!
+          Please create your first amazing plan!
         </Typography>
+        <Button
+          component={Link}
+          to="/account/create"
+          color="inherit"
+          sx={{ marginTop: 3 }}
+        >
+          Create New Plan
+        </Button>
       </Box>
     );
 
@@ -85,58 +93,64 @@ function Bookmarked() {
           marginBottom: 2,
         }}
       >
-        <Typography variant="h4">Bookmarked Plans</Typography>
+        <Typography variant="h4">My Plans</Typography>
+        <Button component={Link} to="/account/create" color="inherit">
+          Create New Plan
+        </Button>
       </Box>
-
       {!isLoading && error && <p>{error}</p>}
-      <Grid container spacing={3}>
+      <Grid container spacing={3} sx={{ width: "100%" }}>
         {isLoading &&
           Array.from({ length: 6 }).map((_, index) => (
-            <Grid size={{ xs: 12, sm: 6, md: 4 }} key={index}>
+            <Grid size={{ xs: 12, sm: 12, md: 6, lg: 4 }} key={index}>
               <PlanCardSkeleton />
             </Grid>
           ))}
         {!isLoading &&
-          plans.map((plan) => {
-            return (
-              <Grid
-                size={{ xs: 12, sm: 6, md: 4 }}
-                key={plan._id}
-                style={{
-                  display: "flex",
-                  flexDirection: "column",
-                  alignItems: "center",
-                }}
-              >
-                <PlanCard
-                  image={plan.images[0]}
-                  title={plan.title}
-                  rate={plan.rate}
-                  type={plan.type}
-                  distance={plan.distance}
-                  stopCount={plan.stopCount}
-                  planId={plan._id}
-                  isBookmarked={true}
-                  showBookmarkBtn={false}
-                />
-                <Box>
-                  <Button
-                    onClick={() => openDeleteDialog(plan._id)}
-                    style={{ backgroundColor: "#f44336", color: "white" }}
-                    sx={{ marginLeft: 1, marginTop: 1 }}
-                  >
-                    Remove
-                  </Button>
-                </Box>
-              </Grid>
-            );
-          })}
+          plans.map((plan) => (
+            <Grid
+              size={{ xs: 12, sm: 12, md: 6, lg: 4 }}
+              key={plan._id}
+              style={{
+                display: "flex",
+                flexDirection: "column",
+                alignItems: "center",
+              }}
+            >
+              <PlanCard
+                image={plan.images[0]}
+                title={plan.title}
+                rate={plan.rate}
+                type={plan.type}
+                distance={plan.distance}
+                stopCount={plan.stopCount}
+                planId={plan._id}
+                showBookmarkBtn={false}
+              />
+              <Box>
+                <Button
+                  component={Link}
+                  to={`/account/${plan._id}`}
+                  sx={{ marginTop: 1 }}
+                >
+                  Edit
+                </Button>
+                <Button
+                  onClick={() => openDeleteDialog(plan._id)}
+                  style={{ backgroundColor: "#f44336", color: "white" }}
+                  sx={{ marginLeft: 1, marginTop: 1 }}
+                >
+                  Remove
+                </Button>
+              </Box>
+            </Grid>
+          ))}
       </Grid>
       <Pagination page={Number(page)} pagesCount={pagesCount} />
       <AlertDialog
         isOpen={alertOpen}
         onClose={handleClose}
-        title="Removing Bookmarked Plan"
+        title="Removing a Plan Permanently"
         message="Are you sure you want to remove this plan?"
         onConfirm={handleRemovePlan}
         confirmText="Remove"
@@ -146,4 +160,4 @@ function Bookmarked() {
   );
 }
 
-export default Bookmarked;
+export default MyPlans;
