+++ src/pages/dashboard/components/StopDialog.jsx
